+++ src/item/MapActorItem.cpp
@@ -146,7 +146,7 @@ void MapActorItem::drawSelectionUI()
                 std::static_pointer_cast<const void>(
                     std::make_shared<const MapActorData>(mSelectionData)
                 ),
-                data_change_flag
+                u32(data_change_flag)
             };
             ActionMgr::instance()->pushAction<ActionItemDataChange>(&context);
         }
